+++ NEWS.md
@@ -4,7 +4,7 @@
 
 ## Package
 
-* Closed #7 by setting up qTrade package using `qCreate::setup_package()`
+* Closed #7 by setting up qTrade package using `manypkgs::setup_package()`
   * Added core package files
   * Added `.github` folder and files
   * Added `tests` folder and files
@@ -20,7 +20,10 @@
 * Closed #4 by adding `LABPTA` dataset to agreements database
 * Cleaned datasets by standardizing titles and dates in databases using `standardise_titles()` and  `standardise_dates()` functions
 * Included variables for agreement type, document type, and listing in WTO for datasets in agreements database
-* Added tests for datasets with `qCreate::export_data()` in agreements database
+* Added tests for datasets with `manypkgs::export_data()` in agreements database
 * Arranged DESTA and GPTAD datasets to reflect treaty membership data by country
-* Added tests for DESTA and GPTAD datasets with `qCreate::export_data()` in memberships database
+* Added tests for DESTA and GPTAD datasets with `manypkgs::export_data()` in memberships database
+
+## Functions
+* Added `code_countryname()` function and tests